refactor(ui): SearchBar를 shared/ui로 통합 및 픽셀 돋보기 아이콘 컴포넌트 추가 (#111)
- shared/ui/SearchBar를 Figma 디자인으로 교체 (h56/rounded8/border #a6a6a6)
- features/search/SearchBar 삭제 + export 제거, 사용처 3곳 import를 shared/ui로 변경
- SearchIcon 컴포넌트 신규 (Figma 픽셀 돋보기, currentColor) + SearchBar에 적용
