+++ src/features/search/index.ts
@@ -1,2 +1,1 @@
-export { SearchBar } from './ui/SearchBar'
 export { PopularKeywords } from './ui/PopularKeywords'
