+++ src/shared/ui/SearchIcon.tsx
@@ -1,0 +1,21 @@
+/** Figma 픽셀 돋보기 아이콘 (icon/ search). 색상은 currentColor로 제어 */
+export const SearchIcon = ({ className }: { className?: string }) => (
+  <svg
+    viewBox="0 0 32 32"
+    fill="none"
+    xmlns="http://www.w3.org/2000/svg"
+    className={className}
+    aria-hidden
+  >
+    <path d="M17.8176 5.99988H10.5449V8.72715H17.8176V5.99988Z" fill="currentColor" />
+    <path d="M10.5456 8.72726H7.81836V11.4545H10.5456V8.72726Z" fill="currentColor" />
+    <path d="M20.5456 8.72726H17.8184V11.4545H20.5456V8.72726Z" fill="currentColor" />
+    <path d="M8.72727 11.4547H6V16.9092H8.72727V11.4547Z" fill="currentColor" />
+    <path d="M22.364 11.4547H19.6367V16.9092H22.364V11.4547Z" fill="currentColor" />
+    <path d="M10.5456 16.9094H7.81836V19.6367H10.5456V16.9094Z" fill="currentColor" />
+    <path d="M20.5455 16.9094H16.9092V20.5458H20.5455V16.9094Z" fill="currentColor" />
+    <path d="M17.8176 19.6363H10.5449V22.3635H17.8176V19.6363Z" fill="currentColor" />
+    <path d="M23.2731 19.6363H19.6367V23.2726H23.2731V19.6363Z" fill="currentColor" />
+    <path d="M25.9996 22.3636H22.3633V26H25.9996V22.3636Z" fill="currentColor" />
+  </svg>
+)
