+++ src/shared/ui/index.ts
@@ -16,6 +16,7 @@ export * from './SectionHeader'
 export * from './Separator'
 export * from './Switch'
 export * from './SearchBar'
+export * from './SearchIcon'
 export * from './Tabs'
 export * from './ListingStats'
 export * from './FavoriteButton'
